requirements.txt
opencv-python
numpy
matplotlib
grpcio
requests
